Refactor: Replace hydra_message_store with telemetry-ingestor
- Removed the hydra_message_store package and its associated files.
- Introduced telemetry-ingestor package with a new implementation for processing and storing messages from the Hydra Gateway.
- Updated Dockerfile and workspace configuration to reflect the new package structure.
- Added new database schema and message handling logic for telemetry data.
