+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
     "sergw",
-    "hydra_message_store",
+    "telemetry-ingestor",
     "hydra_command_dispatcher",
     "hydrate",
     "hydrand",
