+++ Dockerfile
@@ -3,7 +3,7 @@ FROM ubuntu:latest
 WORKDIR /app
 
 
-COPY target/x86_64-unknown-linux-musl/release/hydra_message_store /usr/local/bin/hydra_message_store
+COPY target/x86_64-unknown-linux-musl/release/telemetry-ingestor /usr/local/bin/telemetry-ingestor
 COPY target/x86_64-unknown-linux-musl/release/hydra_command_dispatcher /usr/local/bin/hydra_command_dispatcher
 COPY target/x86_64-unknown-linux-musl/release/hydrate /usr/local/bin/hydrate
 COPY target/x86_64-unknown-linux-musl/release/hydrand /usr/local/bin/hydrand
