+++ db/index.ts
@@ -12,12 +12,13 @@ const noEmptyLines = noComments.replace(/\n\s*\n/g, "");
 
 const statements = noEmptyLines.split(";").map(s => s.trim()).filter(s => s.length > 0);
 
-
+let n_errors = 0;
 for (const statement of statements) {
     try {
         const r = await client.execute(statement);
         console.log(statement.split("\n")[0] + "  ...OK")
     } catch (e) {
+        n_errors++;
         console.error("Statement failed:")
         console.log("--------------------------------")
         console.log(statement)
@@ -26,3 +27,10 @@ for (const statement of statements) {
         console.log("--------------------------------")
     }
 }
+
+if (n_errors > 0) {
+    console.error(`Failed to execute ${n_errors} statements`);
+    process.exit(1);
+}
+
+console.log(`Successfully executed ${statements.length} statements`);
