--- hydra_message_store/.env
@@ -1,1 +1,0 @@
-DATABASE_URL=postgres://postgres:postgres@localhost:5432/postgres
